polish out resizing and reconnecting add video only websockets mode
